修改养老金比率数据类型

+++ yeb-server/src/main/java/com/xxx/server/pojo/Salary.java
@@ -53,7 +53,7 @@ public class Salary implements Serializable {
     private Integer pensionBase;
 
     @ApiModelProperty(value = "养老金比率")
-    private Integer pensionPer;
+    private Float pensionPer;
 
     @ApiModelProperty(value = "启用时间")
     @JsonFormat(pattern = "yyyy-MM-dd",timezone = "Asia/Shanghai")
